rm makefile

+++ build.sh
@@ -6,12 +6,15 @@
 #########################################################################
 #!/bin/bash
 
+pkg_name = $(shell echo CheatServer-`date +%Y%m%d-%H%M%S`.tar.gz)
+echo ${pkg_name}
+
+if [ -d ./build ]; then
 rm -rf build
+fi
+
 mkdir build
 cd build && cmake ..
 make clean
 make -j4
 
-mkdir release
-mkdir release/logs
-mv bin release/
